+++ src/main/kotlin/org/danilopianini/plagiarismdetector/analyzer/TokenizationAnalyzer.kt
@@ -3,9 +3,9 @@ package org.danilopianini.plagiarismdetector.analyzer
 import org.danilopianini.plagiarismdetector.analyzer.representation.TokenizedSource
 import org.danilopianini.plagiarismdetector.analyzer.representation.TokenizedSourceImpl
 import org.danilopianini.plagiarismdetector.analyzer.representation.token.Token
-import org.danilopianini.plagiarismdetector.analyzer.steps.JavaParser
-import org.danilopianini.plagiarismdetector.analyzer.steps.JavaPreprocessor
-import org.danilopianini.plagiarismdetector.analyzer.steps.JavaTokenizer
+import org.danilopianini.plagiarismdetector.analyzer.steps.tokenization.java.JavaParser
+import org.danilopianini.plagiarismdetector.analyzer.steps.tokenization.java.JavaPreprocessor
+import org.danilopianini.plagiarismdetector.analyzer.steps.tokenization.java.JavaTokenizer
 import org.danilopianini.plagiarismdetector.analyzer.steps.StepHandler
 import java.io.File
